+++ src/main/java/org/fabux/discordbotframework/Main.java
@@ -1,0 +1,1 @@
+        VersionHandler.version("fabuxorg", "DiscordBot-Framework");
