+++ src/main/java/org/fabux/discordbotframework/event/core/Disconnect.java
@@ -1,0 +1,11 @@
+package org.fabux.discordbotframework.event.core;
+
+import net.dv8tion.jda.api.events.DisconnectEvent;
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
+
+import javax.annotation.Nonnull;
+
+public class Disconnect extends ListenerAdapter {
+    @Override
+    public void onDisconnect(@Nonnull DisconnectEvent event) {}
+}
