+++ src/main/java/org/fabux/discordbotframework/event/core/Ready.java
@@ -1,0 +1,11 @@
+package org.fabux.discordbotframework.event.core;
+
+import net.dv8tion.jda.api.events.ReadyEvent;
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
+
+import javax.annotation.Nonnull;
+
+public class Ready extends ListenerAdapter {
+    @Override
+    public void onReady(@Nonnull ReadyEvent event) {}
+}
