+++ src/main/java/org/fabux/discordbotframework/event/core/StatusChange.java
@@ -1,0 +1,11 @@
+package org.fabux.discordbotframework.event.core;
+
+import net.dv8tion.jda.api.events.StatusChangeEvent;
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
+
+import javax.annotation.Nonnull;
+
+public class StatusChange extends ListenerAdapter {
+    @Override
+    public void onStatusChange(@Nonnull StatusChangeEvent event) {}
+}
