+++ src/main/java/org/fabux/discordbotframework/event/user/ActivityEnd.java
@@ -1,0 +1,9 @@
+package org.fabux.discordbotframework.event.user;
+
+import net.dv8tion.jda.api.events.user.UserActivityEndEvent;
+
+import javax.annotation.Nonnull;
+
+public class ActivityEnd {
+    public void onUserActivityEnd(@Nonnull UserActivityEndEvent event) {}
+}
