+++ src/main/java/org/fabux/discordbotframework/event/user/ActivityStart.java
@@ -1,0 +1,9 @@
+package org.fabux.discordbotframework.event.user;
+
+import net.dv8tion.jda.api.events.user.UserActivityStartEvent;
+
+import javax.annotation.Nonnull;
+
+public class ActivityStart {
+    public void onUserActivityStart(@Nonnull UserActivityStartEvent event) {}
+}
