+++ src/main/java/org/fabux/discordbotframework/event/user/UserTyping.java
@@ -1,0 +1,10 @@
+package org.fabux.discordbotframework.event.user;
+
+import net.dv8tion.jda.api.events.user.UserTypingEvent;
+
+import javax.annotation.Nonnull;
+
+public class UserTyping {
+    public void onUserTyping(@Nonnull UserTypingEvent event) {}
+
+}
