+++ src/main/java/org/fabux/discordbotframework/event/user/UserUpdateAvatar.java
@@ -1,0 +1,10 @@
+package org.fabux.discordbotframework.event.user;
+
+import net.dv8tion.jda.api.events.user.update.UserUpdateAvatarEvent;
+
+import javax.annotation.Nonnull;
+
+public class UserUpdateAvatar {
+    public void onUserUpdateAvatar(@Nonnull UserUpdateAvatarEvent event) {}
+
+}
