+++ src/main/java/org/fabux/discordbotframework/event/user/UserUpdateName.java
@@ -1,0 +1,9 @@
+package org.fabux.discordbotframework.event.user;
+
+import net.dv8tion.jda.api.events.user.update.UserUpdateNameEvent;
+
+import javax.annotation.Nonnull;
+
+public class UserUpdateName {
+    public void onUserUpdateName(@Nonnull UserUpdateNameEvent event) {}
+}
