+++ src/main/java/org/fabux/discordbotframework/event/user/UserUpdateOnlineStatus.java
@@ -1,0 +1,10 @@
+package org.fabux.discordbotframework.event.user;
+
+import net.dv8tion.jda.api.events.user.update.UserUpdateOnlineStatusEvent;
+
+import javax.annotation.Nonnull;
+
+public class UserUpdateOnlineStatus {
+    public void onUserUpdateOnlineStatus(@Nonnull UserUpdateOnlineStatusEvent event) {}
+
+}
