+++ src/main/java/org/fabux/discordbotframework/plugin/IPlugin.java
@@ -1,0 +1,23 @@
+package org.fabux.discordbotframework.plugin;
+
+import org.jetbrains.annotations.NotNull;
+
+public interface IPlugin {
+    @NotNull
+    static String getName(){return null;}
+
+    @NotNull
+    static Integer getVersion(){return 0;}
+
+    @NotNull
+    static String getDescription(){return null;}
+
+    @NotNull
+    void start();
+
+    @NotNull
+    void stop();
+
+    @NotNull
+    void setPluginManager(PluginManager manager);
+}
