+++ src/main/java/org/fabux/discordbotframework/util/Var.java
@@ -1,0 +1,8 @@
+package org.fabux.discordbotframework.util;
+
+import net.dv8tion.jda.api.JDABuilder;
+
+public class Var {
+    public static JDABuilder builder;
+
+}
