--- src/main/java/wf/key/discordbotframework/Main.java
@@ -1,13 +1,0 @@
-package wf.key.discordbotframework;
-
-import wf.key.discordbotframework.util.versionhandler.VersionHandler;
-
-public class Main {
-    public static void main (String [] args){
-        //Test version
-        VersionHandler.version();
-
-
-
-    }
-}
