--- src/main/java/wf/key/discordbotframework/event/user/ActivityEnd.java
@@ -1,9 +1,0 @@
-package wf.key.discordbotframework.event.user;
-
-import net.dv8tion.jda.api.events.user.UserActivityEndEvent;
-
-import javax.annotation.Nonnull;
-
-public class ActivityEnd {
-    public void onUserActivityEnd(@Nonnull UserActivityEndEvent event) {}
-}
