--- src/main/java/wf/key/discordbotframework/event/user/ActivityStart.java
@@ -1,9 +1,0 @@
-package wf.key.discordbotframework.event.user;
-
-import net.dv8tion.jda.api.events.user.UserActivityStartEvent;
-
-import javax.annotation.Nonnull;
-
-public class ActivityStart {
-    public void onUserActivityStart(@Nonnull UserActivityStartEvent event) {}
-}
