--- src/main/java/wf/key/discordbotframework/event/user/UserTyping.java
@@ -1,10 +1,0 @@
-package wf.key.discordbotframework.event.user;
-
-import net.dv8tion.jda.api.events.user.UserTypingEvent;
-
-import javax.annotation.Nonnull;
-
-public class UserTyping {
-    public void onUserTyping(@Nonnull UserTypingEvent event) {}
-
-}
