--- src/main/java/wf/key/discordbotframework/event/user/UserUpdateAvatar.java
@@ -1,10 +1,0 @@
-package wf.key.discordbotframework.event.user;
-
-import net.dv8tion.jda.api.events.user.update.UserUpdateAvatarEvent;
-
-import javax.annotation.Nonnull;
-
-public class UserUpdateAvatar {
-    public void onUserUpdateAvatar(@Nonnull UserUpdateAvatarEvent event) {}
-
-}
