--- src/main/java/wf/key/discordbotframework/event/user/UserUpdateName.java
@@ -1,9 +1,0 @@
-package wf.key.discordbotframework.event.user;
-
-import net.dv8tion.jda.api.events.user.update.UserUpdateNameEvent;
-
-import javax.annotation.Nonnull;
-
-public class UserUpdateName {
-    public void onUserUpdateName(@Nonnull UserUpdateNameEvent event) {}
-}
