--- src/main/java/wf/key/discordbotframework/event/user/UserUpdateOnlineStatus.java
@@ -1,10 +1,0 @@
-package wf.key.discordbotframework.event.user;
-
-import net.dv8tion.jda.api.events.user.update.UserUpdateOnlineStatusEvent;
-
-import javax.annotation.Nonnull;
-
-public class UserUpdateOnlineStatus {
-    public void onUserUpdateOnlineStatus(@Nonnull UserUpdateOnlineStatusEvent event) {}
-
-}
