--- src/main/java/wf/key/discordbotframework/plugin/IPlugin.java
@@ -1,23 +1,0 @@
-package wf.key.discordbotframework.plugin;
-
-import org.jetbrains.annotations.NotNull;
-
-public interface IPlugin {
-    @NotNull
-    static String getName(){return null;}
-
-    @NotNull
-    static Integer getVersion(){return 0;}
-
-    @NotNull
-    static String getDescription(){return null;}
-
-    @NotNull
-    void start();
-
-    @NotNull
-    void stop();
-
-    @NotNull
-    void setPluginManager(PluginManager manager);
-}
