--- src/main/java/wf/key/discordbotframework/util/Var.java
@@ -1,8 +1,0 @@
-package wf.key.discordbotframework.util;
-
-import net.dv8tion.jda.api.JDABuilder;
-
-public class Var {
-    public static JDABuilder builder;
-
-}
