--- src/main/java/wf/key/discordbotframework/util/initBot/InitBot.java
@@ -1,38 +1,0 @@
-package wf.key.discordbotframework.util.initBot;
-
-import wf.key.discordbotframework.datahandler.filehandler.config.ConfigReader;
-import wf.key.discordbotframework.util.logs.Logger;
-import net.dv8tion.jda.api.AccountType;
-import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.JDABuilder;
-
-import javax.security.auth.login.LoginException;
-
-import static wf.key.discordbotframework.util.Var.builder;
-
-public class InitBot {
-    public InitBot(){
-        builder = new JDABuilder(AccountType.BOT);
-
-        builder.setToken(getToken());
-        builder.setAutoReconnect(true);
-
-        builder.setStatus(HandleOnlineStatus.handleStatus());
-
-        try {
-            JDA jda = builder.build();
-        } catch (LoginException e) {
-            Logger.logger(e.getMessage());
-            e.printStackTrace();
-        }
-    }
-    private String getToken(){
-        if(!ConfigReader.read("data", "config", "discord_token").isEmpty() ||
-        ConfigReader.read("data", "config", "discord_token").length() == 59){
-            System.err.println("Invalid Bot Token");
-        }else {
-            return ConfigReader.read("data", "config", "discord_token");
-        }
-        return null;
-    }
-}
