--- src/main/java/wf/key/discordbotframework/util/logs/Logger.java
@@ -1,9 +1,0 @@
-package wf.key.discordbotframework.util.logs;
-
-import wf.key.discordbotframework.datahandler.filehandler.WriteFile;
-
-public class Logger {
-    public static void logger(String content){
-        WriteFile.writer("logs", "LOG", "log", content);
-    }
-}
